requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Synthetic Data Generation
synthcity==1.0.0
torch==2.1.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Privacy & Security
tensorflow-privacy==0.8.10
diffprivlib==0.6.3
opacus==1.4.0

# Evaluation & Metrics
shap==0.43.0
lime==0.2.0.1
alibi==0.9.4
seaborn==0.13.0
matplotlib==3.8.2

# Medical Imaging (Phase 2)
monai==1.3.0
nibabel==5.1.0
pydicom==2.4.3
opencv-python==4.8.1.78
torchvision==0.16.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Web Interface
streamlit==1.28.2
plotly==5.18.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Docker Support
docker-compose==1.29.2

# System Improvements
progress==1.6
batchprocessing==1.0.0
enhanced-error-handling==1.0.0
model-parameter-customization==1.0.0

# Synthcity Enhancements
synthcity-pate-gan==0.1.0
synthcity-dp-ctgan==0.1.0

# Medical Imaging Enhancements
medical-imaging-frontend-integration==0.1.0
medical-imaging-dicom-support==0.1.0
medical-imaging-preprocessing==0.1.0
medical-imaging-evaluation-metrics==0.1.0

# Additional Documentation and Testing
# Add comprehensive API documentation
# Write more test cases
# Add performance benchmarks
# Document privacy guarantees

# Docker Configuration
# Set up CI/CD pipeline
# Add production configurations
# Implement better error handling